Removed/fixed non-working method for loadCommands() - Changed singleton for main class - Edited references of singleton
